# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "生产就绪的 Go Web 服务脚手架，内置认证、RBAC 权限控制。",
        "title": "Go Server Starter API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/your-username/go-server-starter/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/auth/login/email": {
            "post": {
                "description": "邮箱 + 验证码登录，未注册自动创建账号并绑定 user 角色。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "邮箱验证码登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.AuthLoginByEmailAndCodeReqDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "设备类型: web/desktop/mobile/chromeExtension/api",
                        "name": "Device-Type",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.AuthTokenResDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/login/mobile": {
            "post": {
                "description": "手机号 + 验证码登录，未注册自动创建账号并绑定 user 角色。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "手机验证码登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.AuthLoginByMobileAndCodeReqDto"
                        }
                    },
                    {
                        "type": "string",
                        "description": "设备类型: web/desktop/mobile/chromeExtension/api",
                        "name": "Device-Type",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "租户 ID (默认 default)",
                        "name": "X-Tenant-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.AuthTokenResDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/my-access": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前租户下的角色代码和权限代码，供前端控制入口显示。",
                "tags": [
                    "auth"
                ],
                "summary": "当前用户访问能力",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.MyAccessResDto"
                        }
                    }
                }
            }
        },
        "/auth/send-email-code": {
            "post": {
                "description": "向指定邮箱发送登录验证码。60 秒内不可重复请求。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "description": "邮箱",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.SendEmailCodeReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/send-sms-code": {
            "post": {
                "description": "向指定手机号发送登录验证码。60 秒内不可重复请求。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "发送短信验证码",
                "parameters": [
                    {
                        "description": "手机号",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.SendSmsCodeReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/switch-tenant": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "切换到指定租户，签发新的 JWT token（旧 token 仍有效）。用户必须属于目标租户（默认租户或 user_tenant_refs 成员关系）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "切换租户",
                "parameters": [
                    {
                        "description": "租户 ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.SwitchTenantReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.AuthTokenResDto"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/permission/table": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询可分配的权限代码。",
                "tags": [
                    "permission"
                ],
                "summary": "权限字典",
                "parameters": [
                    {
                        "type": "string",
                        "description": "权限代码",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否启用",
                        "name": "enabled",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-server-starter_internal_dto.PaginationResDto-array_go-server-starter_internal_dto_PermissionResDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-server-starter_internal_dto.AuthLoginByEmailAndCodeReqDto": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "go-server-starter_internal_dto.AuthLoginByMobileAndCodeReqDto": {
            "type": "object",
            "required": [
                "code",
                "countryCode",
                "mobile"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
        "go-server-starter_internal_dto.AuthTokenResDto": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "go-server-starter_internal_dto.MyAccessResDto": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "go-server-starter_internal_dto.PaginationResDto-array_go-server-starter_internal_dto_PermissionResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-server-starter_internal_dto.PermissionResDto"
                    }
                },
                "hasNext": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPage": {
                    "type": "integer"
                }
            }
        },
...
        "go-server-starter_internal_dto.PermissionResDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "go-server-starter_internal_dto.SendEmailCodeReqDto": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "go-server-starter_internal_dto.SendSmsCodeReqDto": {
            "type": "object",
            "required": [
                "countryCode",
                "mobile"
            ],
            "properties": {
                "countryCode": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
        "go-server-starter_internal_dto.SwitchTenantReqDto": {
            "type": "object",
            "required": [
                "tenantId"
            ],
            "properties": {
                "tenantId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\"，登录接口返回的 token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}