# internal/dto/auth_dto.go
package dto

type AuthLoginByMobileAndCodeReqDto struct {
	Mobile      string `json:"mobile" binding:"required"`
	Code        string `json:"code" binding:"required"`
	CountryCode string `json:"countryCode" binding:"required"`
}

type AuthLoginByEmailAndCodeReqDto struct {
	Email string `json:"email" binding:"required,email"`
	Code  string `json:"code" binding:"required"`
}

type AuthTokenResDto struct {
	Token       string   `json:"token"`
	Roles       []string `json:"roles"`
	Permissions []string `json:"permissions"`
}

// SwitchTenantReqDto is the request body for switching to a different tenant.
type SwitchTenantReqDto struct {
	TenantID uint64 `json:"tenantId" binding:"required"`
}